Made total memory a bigger field.

+++ database/migrations/2024_09_13_220218_create_devices_table.php
@@ -23,7 +23,7 @@ class CreateDevicesTable extends Migration
             $table->string('device_name')->nullable();
             $table->string('product_name')->nullable();
             $table->string('supported_cpu_architectures')->nullable();
-            $table->integer('total_memory')->nullable();
+            $table->unsignedBigInteger('total_memory')->nullable();
             $table->integer('device_year_class')->nullable();
             $table->string('manufacturer')->nullable();
             $table->string('model_id')->nullable();
